requirements.txt
asgiref==3.11.0
Django>=5.0
django-modeltranslation==0.19.19
pillow==12.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
